--- src/pageLayouts/LandingLayout/landing.module.scss
@@ -1,171 +1,0 @@
-.landing {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.inner {
-  width: 100%;
-  max-width: 124.8rem;
-  padding: 0 2.4rem;
-
-  @include mobile {
-    padding: 0 1.6rem;
-  }
-}
-
-.bannerImg {
-  height: 55rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: absolute;
-  right: 0;
-  width: 100%;
-  object-fit: cover;
-  z-index: -1;
-
-  @include mobile {
-    height: 24rem;
-  }
-}
-
-.mainTitle {
-  width: 100%;
-  height: 23rem;
-  display: flex;
-  flex-direction: column;
-  margin-top: 15.9rem;
-  gap: 2rem;
-
-  @include tablet {
-    gap: .8rem;
-  }
-
-  @include mobile {
-    height: 9rem;
-    margin-top: 7.4rem;
-  }
-
-  .title {
-    width: 100%;
-    padding-right: 25%;
-    color: var(--white-white_FFFFFF, #fff);
-    font-size: clamp(5.4rem, 5vw, 6.8rem);
-    font-weight: 700;
-    text-overflow: ellipsis;
-    overflow: hidden;
-
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-
-    @include tablet {
-      font-size: clamp(3.6rem, 5vw, 5.4rem);
-    }
-
-    @include mobile {
-      font-size: 2.4rem;
-    }
-  }
-
-  .description {
-    color: var(--white-white_FFFFFF, #fff);
-    font-size: 2.4rem;
-    font-weight: 700;
-
-    @include tablet {
-      font-size: 2rem;
-    }
-
-    @include mobile {
-      font-size: 1.4rem;
-    }
-  }
-}
-
-.searchWrapper {
-  display: flex;
-  width: 100%;
-  margin-top: 10rem;
-  border-radius: 1.6rem;
-  background-color: white;
-  z-index: 3;
-
-  @include tablet {
-    // margin-top: 15.3rem;
-  }
-
-  @include mobile {
-    margin-top: 1.4rem;
-  }
-}
-
-.popularExperiences {
-  width: 100%;
-  margin-top: 4rem;
-  display: flex;
-  flex-direction: column;
-  gap: 3.2rem;
-
-  @include mobile {
-    gap: 1.6rem;
-  }
-
-  .header {
-    display: flex;
-    justify-content: space-between;
-
-    .title {
-      color: var(--black-black_1B1B1B, #1b1b1b);
-      font-size: 3.6rem;
-      font-weight: 700;
-
-      @include mobile {
-        font-size: 1.8rem;
-      }
-    }
-
-    .buttonArrowContainer {
-      display: flex;
-      display: none;
-      gap: 1.2rem;
-
-      @media screen and (min-width: 124.8rem) {
-        display: inline;
-      }
-
-      .buttonArrow {
-        cursor: pointer;
-      }
-    }
-  }
-}
-
-.scroll {
-  width: 100%;
-  overflow: hidden;
-  // 드래그 방지
-  -webkit-user-select: none; /* Safari */
-  -moz-user-select: none; /* Firefox */
-  -ms-user-select: none; /* Internet Explorer/Edge */
-  user-select: none; /* Non-prefixed version, currently supported by Chrome, Edge, Opera and Firefox */
-}
-
-.cardContainer {
-  width: 100%;
-  gap: 2.4rem;
-  display: flex;
-  justify-content: space-around;
-  transition: transform 0.3s ease;
-
-  @include tablet {
-    // gap: 3.2rem;
-  }
-
-  @include mobile {
-    gap: 1.6rem;
-  }
-}
