+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import LandingLayout from '@/pageLayouts/LandingLayout';
+import LandingLayout from '@/pageLayouts/LandingLayout/landingLayout';
 
 export default function Home() {
   return (
